Feeder/requirements.txt
Django
djangorestframework
djangorestframework-xml